+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+- fixed row header width issue for large files
+- fixed issue where large files causes the editor to hang
+	- handsontable virtual renderer is now used with small initial table size
+	- switched from inlining the csv data into the webview html to `webview.postMessage`
+- fixed issue where commands relying on `webview.postMessage` were not working (`edit-csv.apply` and `edit-csv.applyAndSave`)
+
 ## 0.0.10
 
 - fixed issue on windows where fontawesome icons are not loaded